fix/rename template of config file

+++ lifemonitor/api/models/repositories/base.py
@@ -277,14 +277,16 @@ class WorkflowRepository():
             raise IllegalStateException(detail="Missing RO Crate metadata")
         return self.metadata.write_zip(target_path)
 
-    def write(self, target_path: str):
+    def write(self, target_path: str, overwrite: bool = False):
         for f in self.files:
             base_path = os.path.join(target_path, f.dir)
             file_path = os.path.join(base_path, f.name)
-            logger.debug("Writing file: %r", file_path)
             os.makedirs(base_path, exist_ok=True)
-            with open(file_path, "w") as out:
-                out.write(f.get_content())
+            file_exists = os.path.isfile(file_path)
+            if not file_exists or overwrite:
+                logger.debug("%s file: %r", "Overwriting" if file_exists else "Writing", file_path)
+                with open(file_path, "w") as out:
+                    out.write(f.get_content())
 
 
 class IssueCheckResult:
